+++ src/components/overlays/AddServiceProvider/index.tsx
@@ -20,6 +20,7 @@
 
 import {
   Button,
+  CircleProgress,
   DialogActions,
   DialogContent,
   DialogHeader,
@@ -33,19 +34,20 @@ import { useEffect, useState } from 'react'
 import { useDispatch } from 'react-redux'
 import { useTranslation } from 'react-i18next'
 import './style.scss'
-import type { store } from 'features/store'
 import {
   useAddServiceProviderMutation,
   useFetchServiceProviderQuery,
 } from 'features/serviceProvider/serviceProviderApiSlice'
 import Patterns from 'types/Patterns'
 import { setSuccessType } from 'features/serviceProvider/slice'
+import DeleteIcon from '@mui/icons-material/DeleteOutlineOutlined'
 
 export default function AddServiceProvider() {
   const { t } = useTranslation()
-  const dispatch = useDispatch<typeof store.dispatch>()
-  const [inputURL, setInputURL] = useState('')
+  const dispatch = useDispatch()
+  const [inputURL, setInputURL] = useState<string | null>(null)
   const [loading, setLoading] = useState(false)
+  const [deleteLoading, setDeleteLoading] = useState(false)
   const [UrlErrorMsg, setUrlErrorMessage] = useState('')
   const [saveErrorMsg, setSaveErrorMessage] = useState(false)
 
@@ -57,7 +59,8 @@ export default function AddServiceProvider() {
   }, [refetch, dispatch])
 
   const addInputURL = (value: string) => {
-    setInputURL(value)
+    setInputURL(value ?? null)
+    if (!value) return
     if (!Patterns.URL.test(value.trim())) {
       setUrlErrorMessage(t('content.appSubscription.pleaseEnterValidURL'))
     } else {
@@ -66,7 +69,8 @@ export default function AddServiceProvider() {
   }
 
   const addURL = async () => {
-    setLoading(true)
+    if (inputURL) setLoading(true)
+    else setDeleteLoading(true)
     try {
       await addServiceProvider({ url: inputURL }).unwrap()
       dispatch(setSuccessType(true))
@@ -78,7 +82,7 @@ export default function AddServiceProvider() {
   }
 
   return (
-    <>
+    <div className="registerUrlMain">
       <DialogHeader
         {...{
           title: t('content.appSubscription.register.heading'),
@@ -90,29 +94,71 @@ export default function AddServiceProvider() {
 
       <DialogContent>
         <div className="manageInputURL">
-          <Input
-            label={
-              <Typography variant="body2">
-                {t('content.appSubscription.register.endpointConfigured')}
-              </Typography>
-            }
-            value={data ? data.url : ''}
-            disabled={true}
-            sx={{ marginBottom: '20px' }}
-          />
-          <Input
-            label={
-              <Typography variant="body2">
-                {t('content.appSubscription.register.autosetupURL')}
-              </Typography>
-            }
-            placeholder="URL of the customer tentant"
-            onChange={(e) => {
-              addInputURL(e.target.value)
-            }}
-            value={inputURL}
-          />
-          <p className="error">{UrlErrorMsg}</p>
+          {data ? (
+            <>
+              <div className="urlListMain">
+                <Typography variant="body2">
+                  {t('content.appSubscription.register.endpointConfigured')}
+                </Typography>
+                <div className="urlList">
+                  {data?.url ? (
+                    <>
+                      {deleteLoading ? (
+                        <CircleProgress
+                          colorVariant="primary"
+                          interval={800}
+                          iteration
+                          size={20}
+                          step={100}
+                          thickness={1}
+                          variant="indeterminate"
+                        />
+                      ) : (
+                        <DeleteIcon
+                          onClick={() => {
+                            addURL()
+                          }}
+                          className="deleteIcon"
+                        />
+                      )}
+                      <Typography variant="label2" className="urlDetail">
+                        {data.url}
+                      </Typography>
+                    </>
+                  ) : (
+                    '-'
+                  )}
+                </div>
+              </div>
+              <Input
+                label={
+                  <Typography variant="body2">
+                    {t('content.appSubscription.register.autosetupURL')}
+                  </Typography>
+                }
+                placeholder={t(
+                  'content.appSubscription.register.inputPlaceholder'
+                )}
+                onChange={(e) => {
+                  addInputURL(e.target.value)
+                }}
+                value={inputURL}
+              />
+              <p className="error">{UrlErrorMsg}</p>
+            </>
+          ) : (
+            <div className="progress">
+              <CircleProgress
+                colorVariant="primary"
+                interval={800}
+                iteration
+                size={50}
+                step={100}
+                thickness={5}
+                variant="indeterminate"
+              />
+            </div>
+          )}
         </div>
       </DialogContent>
 
@@ -155,6 +201,6 @@ export default function AddServiceProvider() {
         description={t('content.appSubscription.register.providerErrorMessage')}
         showIcon={true}
       />
-    </>
+    </div>
   )
 }