+++ src/components/pages/ServiceReleaseProcess/components/ServiceListOverview.tsx
@@ -20,7 +20,6 @@
 
 import { useState, useEffect, useCallback, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
-import { PageBreadcrumb } from 'components/shared/frame/PageBreadcrumb/PageBreadcrumb'
 import {
   PageHeader,
   ViewSelector,
@@ -30,6 +29,7 @@ import {
   type CardItems,
   PageSnackbar,
   ErrorBar,
+  CircleProgress,
 } from '@catena-x/portal-shared-components'
 import { serviceToCard } from 'features/apps/mapper'
 import { fetchImageWithToken } from 'services/ImageService'
@@ -52,7 +52,7 @@ import {
 import { useDispatch } from 'react-redux'
 import { setServiceReleaseActiveStep } from 'features/serviceManagement/slice'
 import { SuccessErrorType } from 'features/admin/appuserApiSlice'
-import { Box, useTheme, CircularProgress } from '@mui/material'
+import { Box, useTheme } from '@mui/material'
 import { initialState } from 'features/serviceManagement/types'
 
 enum ServiceSubMenuItems {
@@ -201,9 +201,7 @@ export default function ServiceListOverview() {
         title={t('serviceoverview.headerTitle')}
         topPage={true}
         headerHeight={200}
-      >
-        <PageBreadcrumb backButtonVariant="contained" />
-      </PageHeader>
+      />
       <div className="app-main">
         <Box sx={{ marginTop: '20px' }} className="overview-section">
           <section className="overview-section-content">
@@ -229,7 +227,9 @@ export default function ServiceListOverview() {
       <section>
         {isFetching ? (
           <div style={{ textAlign: 'center' }}>
-            <CircularProgress
+            <CircleProgress
+              variant="indeterminate"
+              colorVariant="primary"
               size={50}
               sx={{
                 color: theme.palette.primary.main,