+++ src/components/pages/UsecaseParticipation/index.tsx
@@ -21,10 +21,11 @@ import { useEffect } from 'react'
 import { Link } from 'react-router-dom'
 import { useDispatch } from 'react-redux'
 import { useTranslation, Trans } from 'react-i18next'
-import { CircularProgress, useTheme } from '@mui/material'
+import { useTheme } from '@mui/material'
 import dayjs from 'dayjs'
 import {
   Chip,
+  CircleProgress,
   CustomAccordion,
   PageHeader,
   StatusTag,
@@ -36,7 +37,6 @@ import LaunchIcon from '@mui/icons-material/Launch'
 import uniqueId from 'lodash/uniqueId'
 import { show } from 'features/control/overlay'
 import { OVERLAYS } from 'types/Constants'
-import { PageBreadcrumb } from 'components/shared/frame/PageBreadcrumb/PageBreadcrumb'
 import {
   useFetchUsecaseQuery,
   type UsecaseResponse,
@@ -145,9 +145,7 @@ export default function UsecaseParticipation() {
         title={t('content.usecaseParticipation.headerTitle')}
         topPage={true}
         headerHeight={200}
-      >
-        <PageBreadcrumb backButtonVariant="contained" />
-      </PageHeader>
+      />
       <div className="usecase-main">
         <div className="container">
           <Typography variant="h2" className="heading">
@@ -188,7 +186,9 @@ export default function UsecaseParticipation() {
             <ul>
               {isLoading ? (
                 <div className="progress-main">
-                  <CircularProgress
+                  <CircleProgress
+                    variant="indeterminate"
+                    colorVariant="primary"
                     size={35}
                     sx={{
                       color: theme.palette.primary.main,